Restructure package with uv_build backend
