+++ pyproject.toml
@@ -1,3 +1,7 @@
+[build-system]
+requires = ["uv_build>=0.10.7,<0.11.0"]
+build-backend = "uv_build"
+
 [project]
 name = "ml-runner"
 version = "0.0.1-alpha.1"
