+++ src/ml_runner/configs/aces1.2/luts/ACEScct_to_linear.spi1d
