+++ src/ml_runner/configs/aces1.2/luts/ACESproxy_to_linear.spi1d
