+++ src/ml_runner/configs/aces1.2/luts/linear_to_rec1886.spi1d
