+++ src/ml_runner/configs/aces1.2/luts/linear_to_rec2020.spi1d
