+++ src/ml_runner/configs/sam2.1/sam2.1_hiera_l.yaml
