+++ src/ml_runner/configs/sam2.1_training/sam2.1_hiera_b+_MOSE_finetune.yaml
