+++ src/ml_runner/configs/sam2.1_training/sam21pp_hiera_s.yaml
